refactor(anyharness): split wake coalescing into its own enqueue submodule
Keeps enqueue.rs and its test file under the repo max-lines cap by moving
the coalescing/suppression decision helpers to enqueue/coalescing.rs and
their tests to enqueue/coalescing_tests.rs, mirroring the existing
reconciliation_tests layout.
